pyproject.toml
[tool.poetry]
name = "mike-ehrmantraut-bot"
version = "0.1.0"
description = ""
authors = ["dacoder101 <bobbye23899@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
